+++ app.js
@@ -55,31 +55,9 @@ app.use(cors());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: false}));
 
-app.post('/price_estimate', function(reqs, res) {
-    console.log(reqs.body);
-    var option = {
-        hostname: 'api.uber.com',
-        port: 443,
-        path: '/v1.2/estimates/price?start_latitude=37.7752315&start_longitude=-122.418075&end_latitude=37.7752415&end_longitude=-122.518075&access_token=KA.eyJ2ZXJzaW9uIjoyLCJpZCI6IlNNTmtrVzdVVGtHY1RwUDNUbXY5NUE9PSIsImV4cGlyZXNfYXQiOjE1MjYwMTUwMzAsInBpcGVsaW5lX2tleV9pZCI6Ik1RPT0iLCJwaXBlbGluZV9pZCI6MX0.i_k-F_Qf4YdStWWxTVQ-KFjaQsc4OXvKvMEEajt1wmQ',
-        method: 'get'
-    };
-
-    let req = https.request(option, function (res) {
-        var uberData = '';
-        res.on('data', function (chunk) {
-            uberData = uberData + chunk;
-        });
-        res.on('end', function() {
-            // console.log(uberData);
-            console.log(JSON.parse(uberData));
-            app.render('price_estimate', {resp: '123'});
-        });
-    });
-
-    req.end();
-});
-
-
+// app.post('/price_estimate', function(req, res) {
+//     res.render('home');
+// });
 
 // send app to router
 require('./router')(app);
